requirements.txt
python-telegram-bot==22.0
requests==2.32.3
python-dotenv==1.1.0